Settings: Move show_refresh_rate toggle to Display
