+++ res/xml/development_settings.xml
@@ -229,11 +229,6 @@
             android:fragment="com.android.settings.development.graphicsdriver.GraphicsDriverDashboard" />
 
         <SwitchPreference
-            android:key="show_refresh_rate"
-            android:title="@string/show_refresh_rate"
-            android:summary="@string/show_refresh_rate_summary" />
-
-        <SwitchPreference
             android:key="overlay_settings"
             android:title="@string/overlay_settings_title"
             android:summary="@string/overlay_settings_summary" />
