+++ res/xml/display_settings.xml
@@ -161,6 +161,11 @@
             settings:controller="com.android.settings.display.RefreshRatePreferenceController" />
 
         <SwitchPreference
+            android:key="show_refresh_rate"
+            android:title="@string/show_refresh_rate"
+            android:summary="@string/show_refresh_rate_summary" />
+
+        <SwitchPreference
             android:key="show_operator_name"
             android:title="@string/show_operator_name_title"
             android:summary="@string/show_operator_name_summary"/>
